moving common to common folder
